Code refactoring wip
