+++ src/com/company/DynamicArray.java
@@ -10042,78 +10042,6 @@ class SubSet2 {
         }
     }
 
-    /**
-     * Definition for singly-linked list.
-     * public class ListNode {
-     * int val;
-     * ListNode next;
-     * ListNode() {}
-     * ListNode(int val) { this.val = val; }
-     * ListNode(int val, ListNode next) { this.val = val; this.next = next; }
-     * }
-     */
-    class ReverseLinkedList {
-        public ListNode reverseList(ListNode head) {
-
-            if (head == null) return head;
-            ListNode curr = head, prev = null, next;
-            while (curr != null) {
-                next = curr.next;
-                curr.next = prev;
-                prev = curr;
-                curr = next;
-            }
-            head = prev;
-            return head;
-        }
-    }
-
-    /**
-     * Definition for singly-linked list.
-     * public class ListNode {
-     * int val;
-     * ListNode next;
-     * ListNode() {}
-     * ListNode(int val) { this.val = val; }
-     * ListNode(int val, ListNode next) { this.val = val; this.next = next; }
-     * }
-     */
-    class ReverseLinkedListKGroup {
-        public ListNode reverseKGroup(ListNode head, int k) {
-            if (head == null) return head;
-            int len = getLength(head);
-            return reverse(head, k, len);
-        }
-
-        private int getLength(ListNode curr) {
-            int cnt = 0;
-            while (curr != null) {
-                cnt++;
-                curr = curr.next;
-            }
-            return cnt;
-        }
-
-        // Recursive function to reverse in groups
-        private ListNode reverse(ListNode head, int k, int len) {
-            // base case
-            if (len < k) return head;
-
-            ListNode curr = head, prev = null, next = null;
-            for (int i = 0; i < k; i++) {
-                // reverse linked list code
-                next = curr.next;
-                curr.next = prev;
-                prev = curr;
-                curr = next;
-            }
-
-            ListNode nextNode = reverse(curr, k, len - k);
-            head.next = nextNode;
-            return prev;
-        }
-    }
-
     class MaxTaxiEarningsDP {
 
         // DP + Iteration based solution
